refactor(suite-native): extract receive address components
Refs #29733
