+++ suite-native/module-receive/src/screens/ReceiveAddressScreen.tsx
@@ -1,247 +1,16 @@
-import { useCallback, useEffect } from 'react';
-import { FadeOut } from 'react-native-reanimated';
 import { useSelector } from 'react-redux';
 
-import { G } from '@mobily/ts-belt';
-import {
-    type RouteProp,
-    useNavigation,
-    usePreventRemove,
-    useRoute,
-} from '@react-navigation/native';
+import { type RouteProp, useRoute } from '@react-navigation/native';
 
-import { type DeviceRootState, selectIsDeviceBackupRequired } from '@suite-common/device';
-import { getDisplaySymbol } from '@suite-common/wallet-config';
+import { type DeviceRootState } from '@suite-common/device';
 import {
     type AccountsRootState,
-    selectAccountByKey,
     selectDeviceAccountKeyForNetworkSymbolAndAccountTypeWithIndex,
 } from '@suite-common/wallet-core';
-import { type AccountKey, type TokenAddress } from '@suite-common/wallet-types';
-import { AccountDetailsCard } from '@suite-native/accounts';
-import {
-    AnimatedBox,
-    BoxSkeleton,
-    Card,
-    ErrorMessage,
-    InlineAlertBox,
-    VStack,
-    useBottomSheetModal,
-} from '@suite-native/atoms';
-import {
-    ConfirmOnTrezorWrapper,
-    useConfirmOnTrezorController,
-} from '@suite-native/confirm-on-trezor';
-import { selectHasFirmwareAuthenticityCheckHardFailedForSelectedDevice } from '@suite-native/device';
-import { useInAppRating } from '@suite-native/in-app-rating';
-import { Translation } from '@suite-native/intl';
-import { Link } from '@suite-native/link';
-import { WalletBackupNotSetWarningBottomSheet } from '@suite-native/module-device-onboarding';
-import {
-    type CloseActionType,
-    type ReceiveStackParamList,
-    type ReceiveStackRoutes,
-    Screen,
-} from '@suite-native/navigation';
-import { type TokensRootState, selectAccountTokenSymbol } from '@suite-native/tokens';
-import TrezorConnect from '@trezor/connect';
-import { getScreenWidth } from '@trezor/env-utils';
-import { prepareNativeStyle, useNativeStyles } from '@trezor/styles-native';
-import { HELP_CENTER_OTHER_CRYPTOCURRENCIES_DESTINATION_TAGS_URL } from '@trezor/urls';
+import { type ReceiveStackParamList, type ReceiveStackRoutes } from '@suite-native/navigation';
 
-import { ReceiveBlockedDeviceCompromisedScreen } from './ReceiveBlockedDeviceCompromisedScreen';
-import { ReceiveAddressCard } from '../components/ReceiveAddressCard';
-import { ReceiveScreenHeader } from '../components/ReceiveScreenHeader';
-import { hasReceiveAddressButtonRequest } from '../hooks/receiveSelectors';
-import { useAccountReceiveAddress } from '../hooks/useAccountReceiveAddress';
-
-const SCREEN_WIDTH = getScreenWidth();
-
-const cardStyle = prepareNativeStyle(utils => ({
-    padding: utils.spacings.sp8,
-}));
-
-type ReceiveAddressContentProps = {
-    accountKey: AccountKey;
-    tokenContract?: TokenAddress;
-    closeActionType: CloseActionType;
-};
-
-const ReceiveAddressLoader = ({
-    tokenContract,
-    closeActionType,
-}: Pick<ReceiveAddressContentProps, 'tokenContract' | 'closeActionType'>) => {
-    const { applyStyle } = useNativeStyles();
-
-    return (
-        <Screen
-            header={
-                <ReceiveScreenHeader
-                    tokenContract={tokenContract}
-                    closeActionType={closeActionType}
-                />
-            }
-        >
-            <VStack spacing="sp32" alignItems="center" paddingHorizontal="sp8">
-                <Card style={applyStyle(cardStyle)}>
-                    <BoxSkeleton width={SCREEN_WIDTH - 32} height={70} />
-                </Card>
-                <Card style={applyStyle(cardStyle)}>
-                    <VStack spacing="sp24" alignItems="center" paddingHorizontal="sp24">
-                        <BoxSkeleton width={SCREEN_WIDTH - 80} height={200} />
-                        <BoxSkeleton width={SCREEN_WIDTH - 80} height={48} borderRadius={24} />
-                    </VStack>
-                </Card>
-            </VStack>
-        </Screen>
-    );
-};
-
-const ReceiveAddressContent = ({
-    accountKey,
-    tokenContract,
-    closeActionType,
-}: ReceiveAddressContentProps) => {
-    const { askForRating } = useInAppRating();
-    const account = useSelector((state: AccountsRootState) =>
-        selectAccountByKey(state, accountKey),
-    );
-    const tokenSymbol = useSelector((state: TokensRootState) =>
-        selectAccountTokenSymbol(state, accountKey, tokenContract),
-    );
-    const isDeviceBackupRequired = useSelector(selectIsDeviceBackupRequired);
-
-    const { revealConfirmOnTrezorSheet, confirmOnTrezorRef, closeSheet } =
-        useConfirmOnTrezorController();
-
-    const hasReceiveButtonRequest = useSelector(hasReceiveAddressButtonRequest);
-    const { bottomSheetRef, openModal, closeModal } = useBottomSheetModal();
-
-    const navigation = useNavigation();
-
-    const { address, isReceiveApproved, isUnverifiedAddressRevealed, handleShowAddress } =
-        useAccountReceiveAddress(accountKey);
-
-    usePreventRemove(true, ({ data }) => {
-        TrezorConnect.cancel();
-        navigation.dispatch(data.action);
-        if (isReceiveApproved) {
-            askForRating();
-        }
-    });
-
-    const handleShowReceiveAddress = useCallback(() => {
-        handleShowAddress();
-
-        if (isDeviceBackupRequired) {
-            openModal();
-        }
-    }, [handleShowAddress, openModal, isDeviceBackupRequired]);
-
-    const isLoading =
-        !(isReceiveApproved || hasReceiveButtonRequest) && isUnverifiedAddressRevealed;
-    const isConfirmOnTrezorReady =
-        isUnverifiedAddressRevealed && !isReceiveApproved && hasReceiveButtonRequest;
-
-    useEffect(() => {
-        if (isConfirmOnTrezorReady) revealConfirmOnTrezorSheet();
-    }, [isConfirmOnTrezorReady, revealConfirmOnTrezorSheet]);
-
-    useEffect(() => {
-        if (isReceiveApproved) closeSheet();
-    }, [isReceiveApproved, closeSheet]);
-
-    const closeNoBackupBottomSheet = useCallback(() => {
-        closeModal();
-    }, [closeModal]);
-
-    const hasFirmwareAuthenticityCheckHardFailed = useSelector(
-        selectHasFirmwareAuthenticityCheckHardFailedForSelectedDevice,
-    );
-
-    if (hasFirmwareAuthenticityCheckHardFailed) return <ReceiveBlockedDeviceCompromisedScreen />;
-
-    const isAccountDetailVisible = !isUnverifiedAddressRevealed && !isReceiveApproved;
-
-    if (G.isNullable(account) || G.isNullable(address)) {
-        return <ErrorMessage errorMessage={<Translation id="generic.unknownError" />} />;
-    }
-
-    const onCancel = () => {
-        TrezorConnect.cancel();
-        navigation.goBack();
-    };
-
-    const showDestinationTagInfo =
-        account.networkType === 'ripple' || account.networkType === 'stellar';
-
-    return (
-        <ConfirmOnTrezorWrapper
-            isManualControlEnabled
-            controlRef={confirmOnTrezorRef}
-            closeActionType={closeActionType}
-            closeAction={onCancel}
-            defaultHeader={
-                <ReceiveScreenHeader
-                    accountKey={accountKey}
-                    tokenContract={tokenContract}
-                    closeActionType={isReceiveApproved ? 'close' : closeActionType}
-                />
-            }
-        >
-            <VStack marginTop="sp8" spacing="sp16" flex={1}>
-                {showDestinationTagInfo && (
-                    <InlineAlertBox
-                        intent="info"
-                        title={
-                            <Translation
-                                id="moduleReceive.destinationTag"
-                                values={{
-                                    link: chunk => (
-                                        <Link
-                                            label={chunk}
-                                            textVariant="body-xs"
-                                            href={
-                                                HELP_CENTER_OTHER_CRYPTOCURRENCIES_DESTINATION_TAGS_URL
-                                            }
-                                            isUnderlined
-                                            textColor="contentPrimary"
-                                            textPressedColor="contentSecondary"
-                                        />
-                                    ),
-                                    coinSymbol: tokenSymbol ?? getDisplaySymbol(account.symbol),
-                                }}
-                            />
-                        }
-                    />
-                )}
-                {isAccountDetailVisible && (
-                    <AnimatedBox exiting={FadeOut}>
-                        <AccountDetailsCard accountKey={accountKey} tokenContract={tokenContract} />
-                    </AnimatedBox>
-                )}
-                <ReceiveAddressCard
-                    accountDescriptor={account.descriptor}
-                    symbol={account.symbol}
-                    address={address}
-                    isLoading={isLoading}
-                    deviceStaticSessionId={account.deviceState}
-                    isTokenAddress={!!tokenContract}
-                    isReceiveApproved={isReceiveApproved}
-                    isUnverifiedAddressRevealed={isUnverifiedAddressRevealed}
-                    onShowAddress={handleShowReceiveAddress}
-                />
-            </VStack>
-            {isDeviceBackupRequired && (
-                <WalletBackupNotSetWarningBottomSheet
-                    ref={bottomSheetRef}
-                    onConfirm={closeNoBackupBottomSheet}
-                    onClose={closeModal}
-                />
-            )}
-        </ConfirmOnTrezorWrapper>
-    );
-};
+import { ReceiveAddressContent } from '../components/ReceiveAddressContent';
+import { ReceiveAddressLoader } from '../components/ReceiveAddressLoader';
 
 export const ReceiveAddressScreen = () => {
     const {
@@ -263,7 +32,6 @@ export const ReceiveAddressScreen = () => {
             routeAccountIndex,
         ),
     );
-
     const accountKey = routeAccountKey ?? foundAccountKey;
 
     if (!accountKey) {
